backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pytest==8.3.5
httpx==0.28.1
sqlalchemy==2.0.38
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
redis==5.2.1
